src/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "assurancenet-api"
version = "0.1.0"
description = "AssuranceNet Document Management API - Azure-native UCM replacement"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "gunicorn>=22.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "aioodbc>=0.5.0",
    "alembic>=1.14.0",
    "azure-identity>=1.19.0",
    "azure-storage-blob>=12.24.0",
    "azure-keyvault-secrets>=4.9.0",
    "azure-monitor-opentelemetry>=1.6.4",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "structlog>=24.4.0",
    "pypdf>=5.1.0",
    "python-multipart>=0.0.18",
    "httpx>=0.28.0",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "S",   # flake8-bandit
    "T20", # flake8-print
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "RUF", # ruff-specific
]
ignore = ["S104", "S108"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["../../tests/backend"]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=80"
